Model fidelity: 322 spec-cited fixes across all domains (exhaustive audit)
Every SDK model was audited property-by-property against all 33
authoritative catalog OpenAPI specs (responses, request bodies, nested
objects, array items, enum value sets) after the owner found live getMetros
fields missing from MetroJson. 358 discrepancies were verified with quoted
spec evidence; this commit applies the 322 that are safely fixable, via 11
file-partitioned fix agents + central integration:

- missing response properties added with exact wire names and spec types
  (92 new model classes, e.g. the PrecisionTime ipv4/MD5/PTP family, Port
  package/order/change/interface/LOA models, Stream alert-rule detection
  method, marketplace entitlements/trial, ServiceProfile lastMileConfig)
- create/update builders gained fields users could not send at all:
  service-token name/description/expirationDateTime/project + VD/network
  access-point selectors, Order.termLength, PrecisionTime package/
  connections/ipv4, port bmmrType, NE device vendorConfig, and more
- enums gained missing spec values with forward-compat UNKNOWN fallbacks
  (ConnectionStatus DRAFT/CANCELLED/PENDING_*, PortState PROVISIONED/...,
  LinkProtocolType VXLAN/EVPN_VXLAN, PriceType, PackageType, ChangeStatus)
- misnamed wire mappings corrected (CloudEvent id-vs-uuid, changelog
  aliases); Integer fields that can exceed 2^31 (ASNs, bandwidths) retyped
  to Long; fixtures updated to mirror the SPEC wire shape (incl. the IAM
  AttributeSet object form)

36 findings intentionally held: breaking public-API removals awaiting owner
sign-off (fictional-but-referenced read models like CrossConnect/Shipment/
WorkVisit, extra fields exposed on public interfaces) and spec-vs-live-API
conflicts where the docs spec is likely wrong (NE device fields mirroring
the Terraform provider). Three resources remain unverifiable against any
spec: Fabric PostBulkConnections, port top-statistics, CP Resellers.

Full gate green: 566 default + 934 wiremock + coverage.

+++ src/main/java/api/equinix/javasdk/core/http/response/Pagination.java
@@ -35,11 +35,11 @@ import lombok.Setter;
 @Setter(AccessLevel.PRIVATE)
 public class Pagination {
 
-    private Integer offset;
+    private Long offset;
 
-    private Integer limit;
+    private Long limit;
 
-    private Integer total;
+    private Long total;
 
     private String previous;
 
@@ -56,7 +56,7 @@ public class Pagination {
         if (offset == null || limit == null || limit == 0) {
             return 0;
         }
-        return offset / limit;
+        return (int) (offset / limit);
     }
 
     /**
@@ -66,7 +66,7 @@ public class Pagination {
      * @return the number of results per page
      */
     public int getPageSize() {
-        return (limit != null) ? limit : 0;
+        return (limit != null) ? limit.intValue() : 0;
     }
 
     /**
